pyproject.toml
[project]
name = "easify-seller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "chromadb>=1.2.0",
    "litellm>=1.78.5",
    "openai>=2.5.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
]
